Implement event in course detail

+++ Calendar/src/ch/epfl/calendar/persistence/LocalDatabaseInterface.java
@@ -40,7 +40,7 @@ public interface LocalDatabaseInterface {
      * @param id
      * @return The event that has id
      */
-    Event getEvent(int id);
+    Event getEvent(long id);
 
     /**
      *
